# Remove remove unnecessary base64 encode & decode