--- a/addons/im_livechat/controllers/main.py
+++ b/addons/im_livechat/controllers/main.py
@@ -24,10 +24,9 @@ class LivechatController(http.Controller):
         if isinstance(mock_attachment, list):  # suppose that CSS asset will not required to be split in pages
             mock_attachment = mock_attachment[0]
         # can't use /web/content directly because we don't have attachment ids (attachments must be created)
-        status, headers, content = request.env['ir.http'].binary_content(id=mock_attachment.id, unique=asset.checksum)
-        content_base64 = base64.b64decode(content) if content else ''
-        headers.append(('Content-Length', len(content_base64)))
-        return request.make_response(content_base64, headers)
+        _, headers, content = request.env['ir.http'].binary_content(id=mock_attachment.id, unique=asset.checksum)
+        headers.append(('Content-Length', len(content)))
+        return request.make_response(content, headers)
 
     @http.route('/im_livechat/load_templates', type='json', auth='none', cors="*")
     def load_templates(self, **kwargs):