… odoo/addons/base/tests/test_avatar_mixin.py
… odoo/addons/base/tests/test_image.py
… odoo/addons/base/tests/test_ir_http.py
… odoo/tools/image.py