[FIX] microsoft_calendar: yearly crash
[IMP] Speed Imp: remove unnecessary base64 encode & decode

Avoid to base64 encode, then decode to process assets and images for a ~25% speed improvement.
Change image processing tool to work on images, rather than base64 encoded strings.
   
Renamed some "id" variables into res_id
    
Performance is ~25% faster on assets & images:
    
      /web/assets/...frontend.min.css:    13ms to 7ms,  base64 enc/dec: 2 -> 0
      /web/image/XML_ID:                  10ms to 8ms,  base64 enc/dec: 3 -> 0
      /web/image/res.users/2/avatar_128:  40ms to 31ms, base64 enc/dec: 4 -> 0

--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr