@fpodoo 2022-01-16
@odoo/rd-security Can someone review this branch for merge.