@odony 2022-01-18
@robodoo r- override=ci/style
Preserving API stability and not introducing bugs is more important than "W0622 Redefining built-in 'id'". We don't use that built-in much anyway.

@fpodoo could we please fix the few remaining comments before merging, so we don't have to make another PR for that?