@robodoo 2022-01-18
Staging failed: ci/runbot on dde51387c42819f8d64e521a27f1df84e00bbc85 (view more at http://runbot.odoo.com/runbot/build/12257744)